requirements.txt
#LangChain Core
langchain
langchain-core
langchain-community
langchain-text-splitters
faiss-cpu
pypdf
beautifulsoup4
lxml

#OpenAI Integration
langchain-openai
openai

#Anthropic Integration 
langchain-anthropic 

#Google Gemini (PaLM) Integration 
langchain-google-genai
google-generativeai

#Hugging Face Integration 
langchain-huggingface 
transformers
huggingface-hub 
transformers
torch
sentence-transformers
streamlit
pydantic
email-validator
grandalf

# Environment Variable Management 
python-dotenv

# Machine Learning Utilities 
numpy 
scikit-learn